// frontend/src/types/dashboard.ts
export interface DashboardStats {
  lvl: number;
  exp: number;
  potentialPoints: number;
  str: number;
  dex: number;
  intStat: number;
  con: number;
  coin: number;
  gem: number;
  currentKnightHP: number;
  currentKnightMP: number;
  currentmageHP: number;
  currentMageMP: number;
  currentStamina: number;
}